+++ apps/cli/clixon_cli_api.h
@@ -39,26 +39,6 @@
 #ifndef _CLIXON_CLI_API_H_
 #define _CLIXON_CLI_API_H_
 
-/*
- * Types
- */
-/*! Autocli list keyword type, see clixon-autocli.yang list-keyword-type
- *
- * Assume a YANG LIST: 
- *    list a {
- *       key x;
- *       leaf x;
- *       leaf y;
- *    }
- * Maybe this type should be in cli_autocli.h
- */
-enum autocli_listkw{
-    AUTOCLI_LISTKW_NONE,  /* No extra keywords, only <vars>: a <x> <y> */
-    AUTOCLI_LISTKW_NOKEY, /* Keywords on non-key variables: a <x> y <y> */
-    AUTOCLI_LISTKW_ALL,   /* Keywords on all variables: a x <x> y <y> */
-};
-typedef enum autocli_listkw autocli_listkw_t;
-
 /* 
  * Function Declarations 
  */
@@ -120,6 +100,7 @@ cvec *cvec_append(cvec *cvv0, cvec *cvv1);
 int   cvec_concat_cb(cvec *cvv, cbuf *cb);
 int cli_process_control(clixon_handle h, cvec *vars, cvec *argv);
 int cli_alias_cb(clixon_handle h, cvec *cvv, cvec *argv);
+int cli_cache_clear(clixon_handle h, cvec *cvv, cvec *argv);
 
 /* In cli_show.c */
 int expand_dbvar(void *h, char *name, cvec *cvv, cvec *argv,