+++ include/clixon_custom.h
@@ -296,3 +296,9 @@
  *       unique "config/interface config/subinterface";
  */
 #define YANG_UNIQUE_MULTI_IGNORE
+
+/*! Query backend of autocli cache
+ *
+ * Instead of local dir
+ */
+#define AUTOCLI_CACHE_PROTOCOL