+++ lib/src/clixon_autocli.c
@@ -50,11 +50,17 @@
 /* cligen */
 #include <cligen/cligen.h>
 
-/* libclixon */
-#include <clixon/clixon.h>
-
-#include "clixon_cli_api.h"
-#include "cli_autocli.h"
+/* clixon */
+#include "clixon_queue.h"
+#include "clixon_hash.h"
+#include "clixon_handle.h"
+#include "clixon_string.h"
+#include "clixon_yang.h"
+#include "clixon_xml.h"
+#include "clixon_data.h"
+#include "clixon_map.h"
+#include "clixon_err.h"
+#include "clixon_autocli.h"
 
 /* Mapping from YANG autocli-op to C enum */
 static const map_str2int autocli_op_map[] = {